hf_space/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-football-drive"
version = "0.1.0"
description = "Adversarial NFL play-calling RL environment — LLM offense vs defense"
requires-python = ">=3.10"
dependencies = [
    "openenv-core @ git+https://github.com/meta-pytorch/OpenEnv.git@v0.2.1",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "scikit-learn>=1.3",
    "joblib>=1.3",
    "numpy>=1.24",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["server"]
package-dir = { "server" = "server" }
